extract domain slicing
